исправлен Dao

+++ src/main/java/equant/dao/MessageDaoImpl.java
@@ -7,6 +7,7 @@ import org.hibernate.cfg.Configuration;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 
+import javax.persistence.NoResultException;
 import javax.persistence.TypedQuery;
 
 public class MessageDaoImpl implements MessageDao {
@@ -28,13 +29,24 @@ public class MessageDaoImpl implements MessageDao {
 
     public MessageMod getMinIdAndRemove() {
         EntityManager entityManager = entityManagerFactory.createEntityManager();
-        entityManager.getTransaction().begin();
-        TypedQuery<MessageMod> messageModTypedQuery = (TypedQuery<MessageMod>) entityManager.createQuery(
-                "SELECT  m From MessageMod m WHERE m.id = (SELECT min(m.id) FROM m)");
-        MessageMod messageMod = messageModTypedQuery.getSingleResult();
-        entityManager.remove(messageMod);
-        entityManager.getTransaction().commit();
+        MessageMod messageMod = null;
+        try {
+            entityManager.getTransaction().begin();
+            TypedQuery<MessageMod> messageModTypedQuery = (TypedQuery<MessageMod>) entityManager.createQuery(
+                    "SELECT  m From MessageMod m WHERE m.id = (SELECT min(m.id) FROM m)");
+         messageMod = messageModTypedQuery.getSingleResult();
+            if (messageMod == null) {
+                System.out.println("No message!");
+                return null;
+            }
+            entityManager.remove(messageMod);
+            entityManager.getTransaction().commit();
+        }catch (NoResultException e){
+            new RuntimeException(e);
+        }
+
         entityManager.close();
+
         return messageMod;
     }
 }
